Updated tests to split between tracking and non-tracking for mocking purpose
