+++ src/components/publishing/__test__/article.test.tsx
@@ -1,7 +1,6 @@
-import { mount, shallow } from "enzyme"
+import { shallow } from "enzyme"
 import "jest-styled-components"
 import * as React from "react"
-import Events from "../../../utils/events"
 import Article from "../article"
 import { StandardArticle } from "../fixtures/articles"
 
@@ -10,6 +9,7 @@ jest.mock("react-slick", () => {
   return props => <div>{props.children}</div>
 })
 jest.mock("react-sizeme", () => jest.fn(c => d => d))
+jest.mock("react-tracking", () => jest.fn(c => d => d))
 
 it("indexes and titles images", () => {
   const article = shallow(<Article article={StandardArticle} />)
@@ -18,15 +18,4 @@ it("indexes and titles images", () => {
   expect(article.state("article").sections[4].images[1].index).toBe(1)
   expect(article.state("article").sections[6].images[0].index).toBe(3)
   expect(article.state("article").sections[6].images[1].index).toBe(4)
-})
-
-it("emits analytics events to an event emitter", done => {
-  const article = mount(<Article article={StandardArticle} />)
-  Events.onEvent(data => {
-    expect(data.action).toEqual("share article")
-    done()
-  })
-  const shareUrl = `http://www.artsy.net/article/${StandardArticle.slug}`
-  const fbURL = `https://www.facebook.com/sharer/sharer.php?u=${encodeURIComponent(shareUrl)}`
-  article.find(`[href='${fbURL}']`).simulate("click")
 })
